+++ src/controllers/user-controller.controller.ts
@@ -62,9 +62,12 @@ export class UserControllerController {
         body.username,
         body.password,
       );
+      console.log(userResponse);
       const user: ParseUser = {
+        id: userResponse.objectId,
         username: userResponse.username,
         sessionToken: userResponse.sessionToken,
+        profilePic: userResponse.profilePic?.url,
       };
       return new StandardJsonResponse<Array<ParseUser>>(
         `Logged in as ${user.username}.`,
@@ -74,6 +77,7 @@ export class UserControllerController {
     } catch (e: any) {
       const err: ParseErrorResponse = e;
       this.response.status(err.statusCode);
+      console.log(err);
       return new StandardJsonResponse<null>(
         parseStringifiedJson<ParseError>(err.message).error,
         null,
@@ -162,12 +166,20 @@ export class UserControllerController {
         body.name,
         body.organisation,
       );
+
+      const userProfileResponse = await this.parseService.getProfile(
+        userResponse.sessionToken,
+      );
+
       const user: ParseUser = {
+        id: userResponse.objectId,
         username: userResponse.username,
         sessionToken: userResponse.sessionToken,
+        profilePic: userProfileResponse.profilePic?.url,
       };
+
       return new StandardJsonResponse<Array<ParseUser>>(
-        `Successfully registered as ${user.username}.`,
+        `Successfully registered as ${body.username}.`,
         [user],
         'ParseUser',
       );
@@ -196,11 +208,30 @@ export class UserControllerController {
       const userProfileResponse = await this.parseService.getProfile(
         sessionToken,
       );
+      const userBadgesResponse = await this.parseService.getBadges(
+        sessionToken,
+        userProfileResponse.objectId,
+      );
+      console.log(userBadgesResponse);
       const profile: ParseProfile = {
+        id: userProfileResponse.objectId,
         username: userProfileResponse.username,
         email: userProfileResponse.email,
-        registerDate: new Date(userProfileResponse.createdAt),
+        profilePic: userProfileResponse.profilePic?.url,
+        name: userProfileResponse.name ? userProfileResponse.name : '',
+        organisation: userProfileResponse.organisation
+          ? userProfileResponse.organisation
+          : '',
+        registrationDate: new Date(userProfileResponse.createdAt),
         updatedDate: new Date(userProfileResponse.updatedAt),
+        badges: userBadgesResponse.results.map(badge => {
+          return {
+            title: badge.name,
+            description: badge.description,
+            url: badge.url,
+            image: badge.image.url,
+          };
+        }),
       };
       return new StandardJsonResponse<Array<ParseProfile>>(
         `Fetched profile successfully.`,
@@ -208,6 +239,7 @@ export class UserControllerController {
         'ParseUser',
       );
     } catch (e: any) {
+      console.log(e);
       const err: ParseErrorResponse = e;
       this.response.status(err.statusCode);
       return new StandardJsonResponse<null>(