+++ src/models/index.ts
@@ -39,3 +39,4 @@ export * from './top-20-mn-per-country.model';
 export * from './top-20-mn-per-hhsurvey.model';
 export * from './data-version.model';
 export * from './aggregation-area.model';
+export * from './intervention.model';